e-stat/requirements.txt
jpy-datareader
pandas
SQLAlchemy
psycopg2-binary
python-dotenv